# specs/001-build-a-web/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: CRM Unifier API
  description: Unified CRM Correspondence Platform API
  version: 0.1.0
  contact:
    name: API Support
    email: api@example.com
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
  - url: https://api.crm-unifier.com/v1
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /customers:
    get:
      summary: List customers
      tags: [Customers]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: search
          in: query
          schema:
            type: string
          description: Search by name, email, or phone
      responses:
        '200':
          description: Customer list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerList'

    post:
      summary: Create customer
      tags: [Customers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
  /customers/{customerId}:
    get:
      summary: Get customer details
      tags: [Customers]
      parameters:
        - $ref: '#/components/parameters/CustomerIdParam'
      responses:
        '200':
          description: Customer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          $ref: '#/components/responses/NotFound'
  /customers/{customerId}/timeline:
    get:
      summary: Get customer message timeline
      tags: [Customers, Messages]
      parameters:
        - $ref: '#/components/parameters/CustomerIdParam'
        - name: channel
          in: query
          schema:
            $ref: '#/components/schemas/ChannelType'
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Message timeline
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageTimeline'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CustomerIdParam:
      name: customerId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
    ChannelType:
      type: string
      enum: [sms, email, voice, whatsapp, facebook, instagram]
    MessageDirection:
      type: string
      enum: [inbound, outbound]
...
    Customer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        displayName:
          type: string
        identities:
          type: array
          items:
            $ref: '#/components/schemas/Identity'
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateCustomerRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        displayName:
          type: string
        metadata:
          type: object
    CustomerList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Customer'
        pagination:
          $ref: '#/components/schemas/Pagination'
    Identity:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        type:
          type: string
          enum: [phone, email, social]
        value:
          type: string
        rawValue:
          type: string
        provider:
          type: string
        verified:
          type: boolean
        linkedAt:
          type: string
          format: date-time
...
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        providerMessageId:
          type: string
        customerId:
          type: string
          format: uuid
        conversationId:
          type: string
          format: uuid
        channel:
          $ref: '#/components/schemas/ChannelType'
        direction:
          $ref: '#/components/schemas/MessageDirection'
        from:
          type: string
        to:
          type: string
        timestamp:
          type: string
          format: date-time
        body:
          type: string
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
        providerMeta:
          type: object
        status:
          type: string
          enum: [received, processed, failed]
...
    MessageTimeline:
      type: object
      properties:
        customerId:
          type: string
          format: uuid
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        pagination:
          $ref: '#/components/schemas/Pagination'
    Attachment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        filename:
          type: string
        size:
          type: integer
        url:
          type: string
        thumbnailUrl:
          type: string
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...